refacto actions tests for answers-validation + accordion tests
